Altered the fruit design to be of general consumables

+++ ex4/Pepse/src/pepse/world/Avatar.java
@@ -55,7 +55,7 @@ public class Avatar extends GameObject {
         return energy;
     }
 
-    public void addEnergy(int energy) {
+    public void addEnergy(float energy) {
         if (this.energy + energy > GameConstants.AVATAR_MAX_ENERGY) {
             this.energy = GameConstants.AVATAR_MAX_ENERGY;
             return;
